[Bug-fix] AutoAdmin(Blacklist)

+++ Plugins/AutoAdmin/Model.py
@@ -1,5 +1,5 @@
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, BigInteger
 from Database import engine
 
 base = declarative_base()
@@ -7,6 +7,6 @@ base = declarative_base()
 
 class Blacklist(base):
     __tablename__ = 'Admin_Blacklist'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
 
 base.metadata.create_all(bind=engine)
